changed persons to support :id

+++ Routes/Account/Prss.js
@@ -4,22 +4,31 @@ var ssnUtil = require('../Session.js');
 var router = Express.Router({caseSensitive: true});
 var async = require('async');
 var mongodb = require('mongodb');
+var ObjectId = require('mongodb').ObjectId;
 
 router.baseURL = '/Prss';
 
 
 /* Much nicer versions */
-router.get('/', function(req, res) {
+router.get('/:id', function(req, res) {
 
    var handler = function(err, prsArr) {
-      if (err)
+      if (err) {
          res.status(500).end();
-
-      res.json(prsArr);
+      }
+      else {
+         res.json(prsArr);
+      }
    };
+
+   if (req.params.id !== req.session.id.toString()) {
+      res.status(403).end();
+   }
    // Empty email but is not an admin. Should only return self email
-   req.cnn.collection('User').findOne({_id: req.session.id},
-    {id: 1, email: 1, firstName: 1, lastName: 1}, handler);
+   else {
+      req.cnn.collection('User').findOne({_id: req.session.id},
+       {id: 1, email: 1, firstName: 1, lastName: 1}, handler);
+   }
 });
 
 router.post('/', function(req, res) {
@@ -61,15 +70,15 @@ router.post('/', function(req, res) {
    });
 });
 
-router.put('/', function(req, res) {
+router.put('/:id', function(req, res) {
    var vld = req.validator;
    var body = req.body;
    var cnn = req.cnn;
 
    async.waterfall([
    function (cb) {
-      console.log(req.session.id);
-      if (vld.chain(!body.role, Tags.badValue, ['role'])
+      if (vld.chain(req.params.id === req.session.id.toString(), Tags.noPermission)
+       .chain(!body.role, Tags.badValue, ['role'])
        .chain(!body.hasOwnProperty("email"),
        Tags.forbiddenField, ["email"])
        .chain(!body.hasOwnProperty("whenRegistered"),
@@ -102,15 +111,20 @@ router.put('/', function(req, res) {
 });
 
 /* Only allow deleting of currently logged in user */
-router.delete('/', function(req, res) {
+router.delete('/:id', function(req, res) {
    var vld = req.validator;
 
    async.waterfall([
    function(cb) {
-      req.cnn.collection('User').deleteOne({_id: req.session.id}, cb);
+      console.log(req.params.id === req.session.id.toString());
+      console.log(req.params.id);
+      console.log(req.session.id);
+      if (vld.check(req.params.id === req.session.id.toString(), Tags.noPermission, null, cb)) {
+         req.cnn.collection('User').deleteOne({_id: req.session.id}, cb);
+      }
    },
    function (result, cb) {
-      if (vld.check(result.deletedCount, Tags.notFound) || !err) {
+      if (vld.check(result.deletedCount, Tags.notFound, null, cb)) {
          req.cnn.collection('Recipe').deleteMany({ownerId: req.session.id}, cb);
       }
    },
